requirements.txt
absl-py==1.4.0
accelerate==0.21.0
addict==2.4.0
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
albumentations==1.3.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
basicsr==1.4.2
bcrypt==4.0.1
beautifulsoup4==4.12.2
boltons==23.0.0
boto3==1.29.0
botocore==1.32.0
cachecontrol==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.0
controlnet-aux==0.0.7
cryptography==41.0.3
cssselect2==0.7.0
cycler==0.11.0
cython==3.0.3
dnspython==2.4.2
easydict==1.10
einops==0.6.1
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fairscale==0.4.4
fastapi==0.101.1
filelock==3.12.2
firebase-admin==6.2.0
flatbuffers==23.5.26
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
gdown==4.7.1
gitdb==4.0.10
gitpython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.25.1
huggingface-hub
humanfriendly==10.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
inquirerpy==0.3.4
insightface==0.7.3
iopath==0.1.10
itsdangerous==2.1.2
jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.4
lazy-loader==0.3
lit==16.0.6
lmdb==1.4.1
lxml==4.9.3
markdown==3.5.1
markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mediapipe==0.10.11
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
onefx==0.0.3
onnx==1.14.1
onnxruntime==1.15.1
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
orjson==3.9.10
packaging==23.1
pandas==2.0.3
paramiko==3.3.1
pfzy==0.3.4
piexif==1.1.3
pillow==10.0.0
platformdirs==3.10.0
portalocker==2.7.0
prettytable==3.9.0
prompt-toolkit==3.0.41
proto-plus==1.22.3
protobuf<4,>=3.11
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.21
pydantic==2.3.0
pydantic-core==2.6.3
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pygments==2.17.2
pyjwt==2.8.0
pynacl==1.5.0
pyparsing==3.0.9
pysocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pywavelets==1.4.1
pyyaml==6.0.1
qudida==0.0.4
regex==2023.8.8
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
runpod==1.3.4
s3transfer==0.7.0
safetensors
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.10.1
seaborn==0.12.2
segment-anything==1.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
starlette==0.27.0
svglib==1.5.1
sympy==1.12
tabulate==0.9.0
termcolor==2.3.0
threadpoolctl==3.2.0
timm==0.4.12
tinycss2==1.2.1
tokenizers
tomli==2.0.1
tomlkit==0.12.3
torch==2.0.1
torchaudio==2.0.2
torchsde
torchvision==0.15.2
tqdm==4.66.1
tqdm-loggable==0.1.4
trampoline==0.1.2
triton==2.0.0
typing-extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
ultralytics==8.0.157
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.19.0
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.1
websockets==12.0
werkzeug==3.0.1
xformers==0.0.21
yacs==0.1.8
yapf==0.40.1
yarl==1.9.2
zipp==3.16.2
numba
llama-cpp-python
kornia
pyOpenSSL
zhipuai
openai
clip-interrogator==0.6.0
simple-lama-inpainting
trimesh
diffusers
spandrel
webcolors
blend_modes
wget
transformers==4.46.1